@@ -1,6 +1,6 @@
 {
-  "name": "material-ui-example",
-  "version": "0.8.0",
+  "name": "material-ui-example-browserify",
+  "version": "0.15.0-alpha.2",
   "description": "Sample project that uses material-ui",
   "repository": {
     "type": "git",
@@ -9,22 +9,26 @@
   "scripts": {
     "start": "gulp"
   },
+  "private": true,
   "devDependencies": {
-    "babelify": "^6.1.2",
-    "browser-sync": "^1.8.1",
-    "browserify": "^7.0.3",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babelify": "^7.2.0",
+    "browser-sync": "^2.11.0",
+    "browserify": "^13.0.0",
     "gulp": "^3.8.10",
     "gulp-notify": "^2.1.0",
     "gulp-util": "^3.0.1",
-    "pretty-hrtime": "^0.2.2",
-    "require-dir": "^0.1.0",
+    "pretty-hrtime": "^1.0.1",
+    "require-dir": "^0.3.0",
     "underscore": "^1.7.0",
     "vinyl-source-stream": "^1.0.0",
-    "watchify": "^2.2.1"
+    "watchify": "^3.6.1"
   },
   "dependencies": {
-    "material-ui": ">=0.8",
-    "react": ">=0.13",
-    "react-tap-event-plugin": "^0.1.3"
+    "material-ui": "^0.15.0-alpha.2",
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "react-tap-event-plugin": "^0.2.1"
   }
 }
