@@ -1,6 +1,6 @@
 {
-  "name": "material-ui-example",
-  "version": "0.8.0",
+  "name": "material-ui-example-webpack",
+  "version": "0.15.0-alpha.2",
   "description": "Sample project that uses material-ui",
   "repository": {
     "type": "git",
@@ -10,23 +10,27 @@
     "start": "webpack-dev-server --config webpack-dev-server.config.js --progress --inline --colors",
     "build": "webpack --config webpack-production.config.js --progress --colors"
   },
+  "private": true,
   "devDependencies": {
-    "babel-core": "^5.8.21",
-    "babel-eslint": "^4.0.5",
-    "babel-loader": "^5.3.2",
-    "babel-runtime": "^5.8.25",
-    "eslint": "^1.1.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^3.2.2",
-    "html-webpack-plugin": "^1.6.1",
-    "react-hot-loader": "^1.2.8",
+    "babel-core": "^6.3.26",
+    "babel-eslint": "^6.0.0",
+    "babel-loader": "^6.2.0",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-runtime": "^6.3.19",
+    "eslint": "^2.5.1",
+    "eslint-loader": "^1.3.0",
+    "eslint-plugin-react": "^4.0.0",
+    "html-webpack-plugin": "^2.7.2",
+    "react-hot-loader": "^1.3.0",
     "transfer-webpack-plugin": "^0.1.4",
-    "webpack": "^1.11.0",
-    "webpack-dev-server": "^1.10.1"
+    "webpack": "^1.12.9",
+    "webpack-dev-server": "^1.14.0"
   },
   "dependencies": {
-    "material-ui": ">=0.8",
-    "react": "0.13",
-    "react-tap-event-plugin": "^0.1.3"
+    "material-ui": "^0.15.0-alpha.2",
+    "react": "^0.14.3",
+    "react-dom": "^0.14.3",
+    "react-tap-event-plugin": "^0.2.1"
   }
 }
