@@ -29,54 +36,83 @@
     "url": "https://github.com/callemall/material-ui/issues"
   },
   "homepage": "http://material-ui.com/",
-  "peerDependencies": {
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0",
-    "react-tap-event-plugin": "^0.2.0",
-    "react-addons-transition-group": "^0.14.0",
-    "react-addons-update": "^0.14.0",
-    "react-addons-create-fragment": "^0.14.0",
-    "react-addons-pure-render-mixin": "^0.14.0"
-  },
-  "devDependencies": {
-    "babel": "^5.4.3",
-    "babel-core": "^5.8.21",
-    "babel-eslint": "^4.1.3",
-    "babel-loader": "^5.3.2",
-    "babel-runtime": "^5.8.25",
-    "babelify": "^6.1.3",
-    "browserify-istanbul": "^0.2.1",
-    "chai": "^3.2.0",
-    "eslint": "^1.1.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^3.2.2",
-    "fbjs": "^0.2.1",
-    "gulp": "^3.9.0",
-    "gulp-eslint": "^1.0.0",
-    "html-webpack-plugin": "^1.6.1",
-    "karma": "^0.13.3",
-    "karma-browserify": "^4.2.1",
-    "karma-chai-sinon": "^0.1.5",
-    "karma-coverage": "^0.4.2",
-    "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.0.4",
-    "karma-phantomjs-launcher": "^0.2.0",
-    "karma-webpack": "^1.7.0",
-    "mocha": "^2.2.5",
-    "phantomjs": "^1.9.17",
-    "react": "^0.14.0",
+  "dependencies": {
+    "inline-style-prefixer": "^1.0.2",
+    "keycode": "^2.1.0",
+    "lodash.flowright": "^3.2.1",
+    "lodash.merge": "^4.1.0",
+    "lodash.throttle": "^4.0.0",
     "react-addons-create-fragment": "^0.14.0",
     "react-addons-pure-render-mixin": "^0.14.0",
     "react-addons-transition-group": "^0.14.0",
     "react-addons-update": "^0.14.0",
+    "react-event-listener": "^0.1.1",
+    "recompose": "^0.15.0",
+    "simple-assign": "^0.1.0",
+    "warning": "^2.1.0"
+  },
+  "peerDependencies": {
+    "react": "^0.14.0",
     "react-dom": "^0.14.0",
-    "react-hot-loader": "^1.2.8",
-    "react-router": "^1.0.0-rc1",
+    "react-tap-event-plugin": "^0.2.0"
+  },
+  "files": [
+    "lib",
+    "src"
+  ],
+  "devDependencies": {
+    "app-module-path": "^1.0.5",
+    "babel-cli": "^6.3.17",
+    "babel-core": "^6.3.21",
+    "babel-eslint": "^6.0.0",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-transform-dev-warning": "^0.1.0",
+    "babel-plugin-transform-react-constant-elements": "^6.3.13",
+    "babel-plugin-transform-react-inline-elements": "^6.3.13",
+    "babel-plugin-transform-react-remove-prop-types": "^0.2.2",
+    "babel-plugin-transform-replace-object-assign": "^0.2.1",
+    "babel-polyfill": "^6.3.14",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-react-native": "^1.5.3",
+    "babel-preset-stage-1": "^6.3.13",
+    "babel-runtime": "^6.3.19",
+    "chai": "^3.2.0",
+    "codecov": "^1.0.1",
+    "enzyme": "^2.0.0",
+    "eslint": "^2.5.1",
+    "eslint-plugin-babel": "^3.1.0",
+    "eslint-plugin-material-ui": "./packages/eslint-plugin-material-ui",
+    "eslint-plugin-react": "^4.0.0",
+    "glob": "^7.0.0",
+    "in-publish": "^2.0.0",
+    "isparta": "^4.0.0",
+    "isparta-loader": "^2.0.0",
+    "istanbul": "^0.4.2",
+    "json-loader": "^0.5.4",
+    "karma": "^0.13.3",
+    "karma-chai-sinon": "^0.1.5",
+    "karma-coverage": "^0.5.3",
+    "karma-mocha": "^0.2.0",
+    "karma-mocha-reporter": "^2.0.0",
+    "karma-phantomjs-launcher": "^1.0.0",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^1.7.0",
+    "minimist": "^1.2.0",
+    "mocha": "^2.2.5",
+    "nodemon": "^1.9.1",
+    "phantomjs-prebuilt": "^2.1.6",
+    "react": "^0.14.0",
+    "react-addons-test-utils": "^0.14.7",
+    "react-dom": "^0.14.7",
+    "react-hot-loader": "^1.3.0",
+    "react-look-native": "^1.0.0-beta7",
+    "react-native": "^0.22.2",
     "react-tap-event-plugin": "^0.2.0",
+    "recursive-readdir-sync": "^1.0.6",
     "rimraf": "^2.4.3",
-    "sinon": "^1.15.4",
+    "sinon": "^1.17.3",
     "sinon-chai": "^2.8.0",
-    "transfer-webpack-plugin": "^0.1.4",
-    "webpack": "^1.11.0"
+    "webpack": "^1.12.9"
   }
 }
