@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   iconSize: 24,
 
   desktopGutter: 24,
@@ -8,10 +8,7 @@ module.exports = {
   desktopKeylineIncrement: 64,
   desktopDropDownMenuItemHeight: 32,
   desktopDropDownMenuFontSize: 15,
-  desktopLeftNavMenuItemHeight: 48,
+  desktopDrawerMenuItemHeight: 48,
   desktopSubheaderHeight: 48,
   desktopToolbarHeight: 56,
 };
-
-
-
