@@ -1,10 +1,10 @@
-module.exports = {
+export default {
 
   once(el, type, callback) {
-    let typeArray = type ? type.split(' ') : [];
-    let recursiveFunction = (e) => {
-      e.target.removeEventListener(e.type, recursiveFunction);
-      return callback(e);
+    const typeArray = type ? type.split(' ') : [];
+    const recursiveFunction = (event) => {
+      event.target.removeEventListener(event.type, recursiveFunction);
+      return callback(event);
     };
 
     for (let i = typeArray.length - 1; i >= 0; i--) {
@@ -15,10 +15,9 @@ module.exports = {
   on(el, type, callback) {
     if (el.addEventListener) {
       el.addEventListener(type, callback);
-    }
-    else {
+    } else {
       // IE8+ Support
-      el.attachEvent('on' + type, () => {
+      el.attachEvent(`on${type}`, () => {
         callback.call(el);
       });
     }
@@ -27,18 +26,17 @@ module.exports = {
   off(el, type, callback) {
     if (el.removeEventListener) {
       el.removeEventListener(type, callback);
-    }
-    else {
+    } else {
       // IE8+ Support
-      el.detachEvent('on' + type, callback);
+      el.detachEvent(`on${type}`, callback);
     }
   },
 
-  isKeyboard(e) {
+  isKeyboard(event) {
     return [
       'keydown',
       'keypress',
       'keyup',
-    ].indexOf(e.type) !== -1;
+    ].indexOf(event.type) !== -1;
   },
 };
