@@ -1,14 +1,14 @@
-var webpack = require('webpack');
-var path = require('path');
-var buildPath = path.resolve(__dirname, 'build');
-var nodeModulesPath = path.resolve(__dirname, 'node_modules');
-var TransferWebpackPlugin = require('transfer-webpack-plugin');
+const webpack = require('webpack');
+const path = require('path');
+const buildPath = path.resolve(__dirname, 'build');
+const nodeModulesPath = path.resolve(__dirname, 'node_modules');
+const TransferWebpackPlugin = require('transfer-webpack-plugin');
 
-var config = {
-  entry: [path.join(__dirname, '/src/app/app.jsx')],
+const config = {
+  entry: [path.join(__dirname, '/src/app/app.js')],
   resolve: {
     //When require, do not have to add these extensions to file's name
-    extensions: ["", ".js", ".jsx"]
+    extensions: ["", ".js"],
     //node_modules: ["web_modules", "node_modules"]  (Default Settings)
   },
   //Render source-map file for final build
@@ -16,43 +16,43 @@ var config = {
   //output config
   output: {
     path: buildPath,    //Path of output file
-    filename: 'app.js'  //Name of output file
+    filename: 'app.js',  //Name of output file
   },
   plugins: [
     //Minify the bundle
     new webpack.optimize.UglifyJsPlugin({
       compress: {
         //supresses warnings, usually from module minification
-        warnings: false
-      }
+        warnings: false,
+      },
     }),
     //Allows error warnings but does not stop compiling. Will remove when eslint is added
     new webpack.NoErrorsPlugin(),
     //Transfer Files
     new TransferWebpackPlugin([
-      {from: 'www'}
-    ], path.resolve(__dirname,"src"))
+      {from: 'www'},
+    ], path.resolve(__dirname,"src")),
   ],
   module: {
     preLoaders: [
       {
-        test: /\.(js|jsx)$/,
+        test: /\.js$/,
         loader: 'eslint-loader',
         include: [path.resolve(__dirname, "src/app")],
-        exclude: [nodeModulesPath]
+        exclude: [nodeModulesPath],
       },
     ],
     loaders: [
       {
-        test: /\.(js|jsx)$/, //All .js and .jsx files
-        loader: 'babel-loader?optional=runtime&stage=0', //react-hot is like browser sync and babel loads jsx and es6-7
-        exclude: [nodeModulesPath]
-      }
-    ]
+        test: /\.js$/, // All .js files
+        loaders: ['babel'], //react-hot is like browser sync and babel loads jsx and es6-7
+        exclude: [nodeModulesPath],
+      },
+    ],
   },
   //Eslint config
   eslint: {
-    configFile: '.eslintrc' //Rules for eslint
+    configFile: '.eslintrc', //Rules for eslint
   },
 };
 
