@@ -1,11 +1,13 @@
-module.exports = {
-  AutoPrefix: require('./auto-prefix'),
-  Colors: require('./colors'),
-  Spacing: require('./spacing'),
-  ThemeManager: require('./theme-manager'),
-  Transitions: require('./transitions'),
-  Typography: require('./typography'),
-  LightRawTheme: require('./raw-themes/light-raw-theme'),
-  DarkRawTheme: require('./raw-themes/dark-raw-theme'),
-  ThemeDecorator: require('./theme-decorator'),
-};
+export MuiThemeProvider from './MuiThemeProvider';
+export * as colors from './colors';
+export darkBaseTheme from './baseThemes/darkBaseTheme';
+export DarkRawTheme from './baseThemes/darkBaseTheme';
+export lightBaseTheme from './baseThemes/lightBaseTheme';
+export LightRawTheme from './baseThemes/lightBaseTheme';
+export getMuiTheme from './getMuiTheme';
+export muiThemeable from './muiThemeable';
+export spacing from './spacing';
+export themeManager from './themeManager';
+export transitions from './transitions';
+export typography from './typography';
+export zIndex from './zIndex';
